Split editor functionality
